pagination is output to the component
