+++ src/components/Header/header.style.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-export const Header = styled.div`
+export const Header = styled.header`
     position: sticky;
     top: 0px;
     left: 0;
