+++ src/pages/Layout/Layout.jsx
@@ -2,17 +2,22 @@ import { Outlet } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import * as S from './layout.style';
 import Header from '../../components/Header/Header';
-import { setIsOpenSort } from '../../store/slices/users';
+import { setIsOpenSort, setIsOpenDataAmount } from '../../store/slices/users';
 
 export default function Layout() {
     const dispatch = useDispatch();
-    const isOpenSort = useSelector((state) => state.users.isOpenSort);
+    const { isOpenSort, isOpenDataAmount } = useSelector(
+        (state) => state.users,
+    );
 
     return (
         <S.Wrapper
             onClick={() => {
                 if (isOpenSort) {
                     dispatch(setIsOpenSort(false));
+                }
+                if (isOpenDataAmount) {
+                    dispatch(setIsOpenDataAmount(false));
                 }
             }}
         >
